Game now can handle when player runs out of fuel and when all players are out of fuel
As in title players can now run out of fuel and their location will not be updated, furthermore if the players are all out of fuel the game is over with no winners. Some adjustments have also been made to the end game text so that all the players positions are displayed before the game ending message. Aswell as a small fix to inconsistency between the play again text.